% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Descriptive Statistics on variables within a data frame}
\usage{
descriptives(data, vars, groupby = NULL, conf.level = 0.95,
  medianCI = FALSE)
}
\arguments{
\item{data}{a \code{\link{data.frame}} containing columns of variables.}

\item{vars}{a \code{\link{vector}} of variable names existing in the dataframe to perform descriptive statistics on.}

\item{groupby}{an optional \code{\link{vector}} containing factor names within a dataframe used for by group processing.}

\item{conf.level}{the confidence level for mean and median confidence intervals.}

\item{medianCI}{a TRUE/FALSE boolean value indicating whether or not to produce bootstrapped confidence intervals on the median (computationally intensive).}
}
\value{
A \code{\link{list}} containing descriptive statistics for each variable listed in the \code{vars} statement.
}
\description{
The  \code{descriptives} function is used to perform descriptive statistics on a set of variables in a \code{\link{data.frame}}.
This can include descriptive statistics by group when including a vector of categorical variables in the groupby= argument.
}
\examples{
descriptives(data=mtcars, vars=c("mpg","disp"), groupby=c("gear","am"))
}
