% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report}
\alias{report}
\title{Produce Output using Pandoc tables}
\usage{
report(object, style = "multiline", round = 4)
}
\arguments{
\item{object}{An R object from an aca analysis.}

\item{style}{a pandoc table style: \code{simple, multiline, grid or rmarkdown}.}

\item{round}{the number of digits to round too when producing the report.}
}
\value{
a printed pander table
}
\description{
The \code{report} function will produce ASCII formated output. This includes both custom and pandoc tables and figures.
}
\examples{
out<-descriptives(data=mtcars, vars=c("mpg","disp"), groupby=c("gear","am"))
report(out)
}
