% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordreports.R
\name{write.report}
\alias{write.report}
\title{Write Formatted Output to a *.docx File}
\usage{
write.report(object, path, round = NULL)
}
\arguments{
\item{object}{An R object from an aca analysis.}

\item{path}{Path or connection to write to.}

\item{round}{round the number of digits to round too when producing the report.}
}
\value{
will produce an *.docx file.
}
\description{
The write.report function will produce an MS Word document using the \code{\link{officer}}.
}
\examples{
out<-descriptives(data=mtcars, vars=c("mpg","disp"), groupby=c("gear","am"))
write.report(out, path="C:/Temp/Output.docx")
}
